last improvements
